Namespaces unit tests.
