+++ Hestia/Hestia.UnitTests/backend/models/ActivatorTests.cs
@@ -6,7 +6,7 @@ using Moq;
 using Newtonsoft.Json.Linq;
 using System.Collections.Generic;
 
-namespace Hestia.UnitTests.backend.models
+namespace Hestia.UnitTests.Backend.Models
 {
     [TestClass]
     public class ActivatorTests
