+++ Hestia/Hestia.UnitTests/backend/models/deserializers/DeviceDeserializerTests.cs
@@ -5,7 +5,7 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Newtonsoft.Json.Linq;
 using System.Collections.Generic;
 
-namespace Hestia.UnitTests.backend.models.deserializers
+namespace Hestia.UnitTests.Backend.Models.Deserializers
 {
     [TestClass]
     public class DeviceDeserializerTests
