+++ Hestia/Hestia.UnitTests/backend/utils/PingServerTests.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace Hestia.UnitTests.backend.utils
+namespace Hestia.UnitTests.Backend.Utils
 {
     [TestClass]
     public class PingServerTests
